requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==2.6.0               # via kombu
astroid==2.4.2            # via pylint
attrs==19.3.0             # via -r requirements.in, jsonschema
billiard==3.6.3.0         # via celery
celery==4.4.6             # via -r requirements.in
click==7.1.2              # via flask
environs==7.4.0           # via -r requirements.in
fire==0.3.1               # via -r requirements.in
flask==1.1.2              # via -r requirements.in
future==0.18.2            # via celery
gunicorn==20.0.4          # via -r requirements.in
isort==4.3.21             # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
jsonschema==3.2.0         # via -r requirements.in
kombu==4.6.11             # via celery
lazy-object-proxy==1.4.3  # via astroid
markupsafe==1.1.1         # via jinja2
marshmallow==3.7.0        # via environs
mccabe==0.6.1             # via pylint
pika==1.1.0               # via -r requirements.in
psycopg2-binary==2.8.5    # via -r requirements.in
pylint==2.5.0             # via -r requirements.in
pyrsistent==0.16.0        # via jsonschema
python-dotenv==0.14.0     # via environs
pytz==2020.1              # via celery
six==1.15.0               # via astroid, fire, jsonschema, pyrsistent
sqlalchemy==1.3.18        # via -r requirements.in
termcolor==1.1.0          # via fire
toml==0.10.1              # via pylint
vine==1.3.0               # via amqp, celery
werkzeug==1.0.1           # via flask
wrapt==1.12.1             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
